float to unsigned int error

+++ image.h
@@ -11,8 +11,8 @@ typedef struct
 typedef struct
 {
 	pixel_t *pixels;
-	size_t width;
-	size_t height;
+	uint16_t width;
+	uint16_t height;
 } bitmap_t;
 
 extern bitmap_t image;
